requirements.txt
selenium==4.18.1
webdriver-manager==4.0.1
beautifulsoup4==4.12.3
python-telegram-bot==20.8
apscheduler==3.10.4
PyYAML==6.0.1
